# openapi.yml
openapi: 3.1.0
info:
  description: The same-origin contract consumed by the BeyondPilot web application.
  title: BeyondPilot API
  version: 1.0.0
servers:
- description: Same origin as the web application
  url: /
...
paths:
...
  /api/solution/admin/solutions:
    get:
      operationId: listAdminSolutions
      parameters:
      - description: "Solutions whose name contains this, ignoring case."
        in: query
        name: q
        required: false
        schema:
          type:
          - string
          - "null"
          maxLength: 100
          minLength: 0
      - description: Only solutions of this status. Drafts are never listed.
        in: query
        name: status
        required: false
        schema:
          type: string
          enum:
          - submitted
          - approved
          - rejected
          pattern: submitted|approved|rejected
      - description: "The page, counted from 1."
        in: query
        name: page
        required: false
        schema:
          type: integer
          format: int32
          default: 1
          minimum: 1
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminSolutionList"
          description: One page of the solutions the parameters select.
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: A parameter is not valid.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The caller is not an operator.
      security:
      - session: []
      summary: "The submitted solutions, those waiting for review first"
      tags:
      - Solution administration
  /api/solution/admin/solutions/{id}:
    get:
      operationId: getAdminSolution
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Solution"
          description: The solution.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The caller is not an operator.
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: There is no such submitted solution.
      security:
      - session: []
      summary: One solution as an operator reviews it
      tags:
      - Solution administration
  /api/solution/admin/solutions/{id}/approve:
    post:
      operationId: approveSolution
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      responses:
        "204":
          description: The solution is approved.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The caller is not an operator.
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: There is no such submitted solution.
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The solution is not waiting for review.
      security:
      - session: []
      summary: Approve a solution that waits for review
      tags:
      - Solution administration
  /api/solution/admin/solutions/{id}/reject:
    post:
      operationId: rejectSolution
      parameters:
      - in: path
        name: id
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RejectSolution"
        required: true
      responses:
        "204":
          description: "The solution is rejected, with the reason."
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: A member is not valid.
        "401":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: Nobody is signed in.
        "403":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The caller is not an operator.
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: There is no such submitted solution.
        "409":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: The solution is neither waiting for review nor approved.
      security:
      - session: []
      summary: "Reject a solution that waits for review, or take an approved one out\
        \ of the directory"
      tags:
      - Solution administration
  /api/solution/deployments:
    get:
      operationId: listCustomerDeployments
      parameters:
      - description: The address of the organization's public page.
        in: query
        name: organization
        required: true
        schema:
          type: string
          maxLength: 120
          minLength: 0
      - description: "The page, from 1."
        in: query
        name: page
        required: false
        schema:
          type:
          - integer
          - "null"
          format: int32
          maximum: 1000
          minimum: 1
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicCustomerDeploymentList"
          description: One page of the deployments.
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: A parameter is not valid.
      summary: "The approved customer deployments of an organization, the most recently\
        \ approved first"
      tags:
      - Solutions
...
  /api/solution/solutions:
    get:
      operationId: listSolutions
      parameters:
      - description: "Solutions whose name or summary contains this, ignoring case."
        in: query
        name: q
        required: false
        schema:
          type:
          - string
          - "null"
          maxLength: 100
          minLength: 0
      - description: Only solutions for this industry.
        in: query
        name: industry
        required: false
        schema:
          type:
          - string
          - "null"
          pattern: banking_finance|insurance|retail_ecommerce|manufacturing|logistics|healthcare|education|real_estate|telecom|energy|agriculture|travel_hospitality|media_entertainment|public_sector|professional_services|technology|automotive_mobility|consumer_goods|other
      - description: Only solutions of this focus area.
        in: query
        name: focusArea
        required: false
        schema:
          type:
          - string
          - "null"
          pattern: conversational_ai|document_processing|computer_vision|speech_voice|predictive_analytics|recommendation|process_automation|ai_agents|generative_content|search_knowledge|data_platform|ai_security|other
      - description: Only solutions of this maturity.
        in: query
        name: maturity
        required: false
        schema:
          type:
          - string
          - "null"
          pattern: idea|prototype|pilot|production|scaled
      - description: Only solutions of the organization at this address.
        in: query
        name: organization
        required: false
        schema:
          type:
          - string
          - "null"
          maxLength: 120
          minLength: 0
      - description: "The order: by name, or the most recently approved first."
        in: query
        name: sort
        required: false
        schema:
          type: string
          default: name
          enum:
          - name
          - newest
          pattern: name|newest
      - description: "The page, counted from 1."
        in: query
        name: page
        required: false
        schema:
          type: integer
          format: int32
          default: 1
          minimum: 1
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicSolutionList"
          description: One page of the solutions the parameters select.
        "400":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: A parameter is not valid.
      summary: "The approved, listed solutions, in the order asked for"
      tags:
      - Solutions
  /api/solution/solutions/{slug}:
    get:
      operationId: getSolution
      parameters:
      - in: path
        name: slug
        required: true
        schema:
          type: string
      responses:
        "200":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublicSolution"
          description: The solution.
        "404":
          content:
            application/problem+json:
              schema:
                $ref: "#/components/schemas/Problem"
          description: "No approved, listed solution has this address."
      summary: One solution of the directory by its address
      tags:
      - Solutions
...
components:
  schemas:
...
    AdminSolutionList:
      type: object
      description: "One page of submitted solutions: those waiting for review first,\
        \ the longest wait on top."
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/SolutionSummary"
        page:
          type: integer
          format: int32
          description: "The page returned, counted from 1."
        pageSize:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
          description: "How many solutions match, over all pages."
      required:
      - items
      - page
      - pageSize
      - total
...
    CustomerDeployment:
      type: object
      description: "A customer deployment of a solution as its organization, and operators,\
        \ see it."
      properties:
        channels:
          type:
          - string
          - "null"
        customer:
          type: string
        decisionMessage:
          type:
          - string
          - "null"
          description: What the operator wrote to the owners with the rejection.
        decisionReason:
          type:
          - string
          - "null"
          description: Why it was last rejected.
          enum:
          - incomplete
          - unverifiable
          - other
        delivered:
          type: string
        id:
          type: string
          format: uuid
        languages:
          type:
          - string
          - "null"
        period:
          type:
          - string
          - "null"
        problem:
          type: string
        result:
          type:
          - string
          - "null"
        stage:
          type: string
          enum:
          - pilot
          - production
        status:
          type: string
          enum:
          - submitted
          - approved
          - rejected
        title:
          type: string
        updatedAt:
          type: string
          format: date-time
        version:
          type: integer
          format: int64
          description: "Sent back with a save, which is refused when the deployment\
            \ changed since."
      required:
      - customer
      - delivered
      - id
      - problem
      - stage
      - status
      - title
      - updatedAt
      - version
...
    Problem:
      type: object
      additionalProperties: false
      properties:
        code:
          type: string
          description: Stable failure code for module failures and request validation;
            clients branch on it.
          maxLength: 63
          pattern: "^[A-Z][A-Z0-9_]+[A-Z0-9]$"
        detail:
          type: string
          description: Safe human-readable explanation; never branch on it.
        errors:
          type: array
          description: "Request validation only: one entry per violation. A rejected\
            \ value is never echoed."
          items:
            type: object
            additionalProperties: false
            properties:
              code:
                type: string
                description: "Constraint code, for example `NOT_BLANK` or `SIZE`."
                pattern: "^[A-Z][A-Z0-9_]+[A-Z0-9]$"
              detail:
                type: string
                description: Safe fallback text; clients translate by code.
              params:
                type: object
                additionalProperties: false
                description: Numeric bounds of the constraint; omitted when it has
                  none.
                properties:
                  max:
                    type: number
                  min:
                    type: number
              pointer:
                type: string
                description: "JSON Pointer to the offending member, for example `#/title`."
            required:
            - code
            - detail
            - pointer
        instance:
          type: string
          format: uri-reference
          description: Request path that produced the problem.
        requestId:
          type: string
          format: uuid
          description: Identifier of the request in the server logs and the `X-Request-Id`
            header.
        status:
          type: integer
          format: int32
          description: HTTP status code.
        title:
          type: string
          description: Short summary of the problem kind; never branch on it.
        type:
          type: string
          format: uri
          description: '`urn:beyondpilot:failure:<code in kebab case>` when `code`
            is present; omitted means `about:blank`.'
      required:
      - requestId
      - status
      - title
    PublicCustomerDeployment:
      type: object
      description: An approved customer deployment as anyone reads it.
      properties:
        approvedAt:
          type: string
          format: date-time
          description: When GenAI Fund approved it.
        channels:
          type:
          - string
          - "null"
        customer:
          type: string
        delivered:
          type: string
        id:
          type: string
          format: uuid
        languages:
          type:
          - string
          - "null"
        period:
          type:
          - string
          - "null"
        problem:
          type: string
        result:
          type:
          - string
          - "null"
        solutionName:
          type: string
        solutionSlug:
          type: string
        stage:
          type: string
          enum:
          - pilot
          - production
        title:
          type: string
      required:
      - approvedAt
      - customer
      - delivered
      - id
      - problem
      - solutionName
      - solutionSlug
      - stage
      - title
    PublicCustomerDeploymentList:
      type: object
      description: One page of the customer deployments of an organization.
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PublicCustomerDeployment"
        page:
          type: integer
          format: int32
        pageSize:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
      required:
      - items
      - page
      - pageSize
      - total
...
    PublicSolution:
      type: object
      description: "An approved, listed solution as anyone reads it."
      properties:
        country:
          type:
          - string
          - "null"
          description: ISO 3166-1 alpha-2.
        customerDeployments:
          type: array
          description: "Its approved customer deployments, the most recently approved\
            \ first."
          items:
            $ref: "#/components/schemas/PublicCustomerDeployment"
        deployment:
          type: array
          items:
            type: string
        focusAreas:
          type: array
          items:
            type: string
        industries:
          type: array
          items:
            type: string
        maturity:
          type:
          - string
          - "null"
          enum:
          - idea
          - prototype
          - pilot
          - production
          - scaled
        name:
          type: string
        organizationName:
          type: string
        organizationSlug:
          type: string
          description: The address of the organization's public page.
        problemsSolved:
          type:
          - string
          - "null"
        slug:
          type: string
        summary:
          type:
          - string
          - "null"
        valueProposition:
          type:
          - string
          - "null"
        website:
          type:
          - string
          - "null"
      required:
      - customerDeployments
      - deployment
      - focusAreas
      - industries
      - name
      - organizationName
      - organizationSlug
      - slug
    PublicSolutionList:
      type: object
      description: "One page of the public directory of solutions, by name."
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/PublicSolutionSummary"
        page:
          type: integer
          format: int32
          description: "The page returned, counted from 1."
        pageSize:
          type: integer
          format: int32
        total:
          type: integer
          format: int64
          description: "How many solutions match, over all pages."
      required:
      - items
      - page
      - pageSize
      - total
    PublicSolutionSummary:
      type: object
      description: One solution in the public directory.
      properties:
        country:
          type:
          - string
          - "null"
          description: ISO 3166-1 alpha-2.
        customerDeployments:
          type: integer
          format: int32
          description: How many approved customer deployments it lists.
        focusAreas:
          type: array
          items:
            type: string
        industries:
          type: array
          items:
            type: string
        maturity:
          type:
          - string
          - "null"
          enum:
          - idea
          - prototype
          - pilot
          - production
          - scaled
        name:
          type: string
        organizationName:
          type: string
        organizationSlug:
          type: string
          description: The address of the organization's public page.
        slug:
          type: string
        summary:
          type:
          - string
          - "null"
      required:
      - customerDeployments
      - focusAreas
      - industries
      - name
      - organizationName
      - organizationSlug
      - slug
...
    RejectSolution:
      type: object
      description: "Why a solution is not approved, and what its owners are told."
      properties:
        message:
          type:
          - string
          - "null"
          description: Shown to the owners with the rejection.
          maxLength: 1000
          minLength: 0
        reason:
          type: string
          enum:
          - incomplete
          - not_an_ai_solution
          - duplicate
          - unverifiable
          - other
          pattern: incomplete|not_an_ai_solution|duplicate|unverifiable|other
      required:
      - reason
...
    Solution:
      type: object
      description: "A solution as its organization, and operators, see it."
      properties:
        complete:
          type: boolean
          description: "Whether it has what a submission needs: a summary, a maturity,\
            \ a focus area and an industry."
        customerDeployments:
          type: array
          description: "Its customer deployments, the newest first, whatever their\
            \ review says."
          items:
            $ref: "#/components/schemas/CustomerDeployment"
        decisionMessage:
          type:
          - string
          - "null"
          description: What the operator wrote to the owners with the rejection.
        decisionReason:
          type:
          - string
          - "null"
          description: Why it was last rejected.
          enum:
          - incomplete
          - not_an_ai_solution
          - duplicate
          - unverifiable
          - other
        deployment:
          type: array
          items:
            type: string
        focusAreas:
          type: array
          items:
            type: string
        id:
          type: string
          format: uuid
        industries:
          type: array
          items:
            type: string
        listed:
          type: boolean
          description: Whether it appears in the public directory once approved.
        maturity:
          type:
          - string
          - "null"
          description: Null is unknown.
          enum:
          - idea
          - prototype
          - pilot
          - production
          - scaled
        name:
          type: string
        organizationId:
          type: string
          format: uuid
        organizationName:
          type: string
        problemsSolved:
          type:
          - string
          - "null"
        slug:
          type: string
        status:
          type: string
          enum:
          - draft
          - submitted
          - approved
          - rejected
        submittedAt:
          type:
          - string
          - "null"
          format: date-time
        summary:
          type:
          - string
          - "null"
        updatedAt:
          type: string
          format: date-time
        valueProposition:
          type:
          - string
          - "null"
        version:
          type: integer
          format: int64
          description: "Sent back with a save, which is refused when the solution\
            \ changed since."
        website:
          type:
          - string
          - "null"
      required:
      - complete
      - customerDeployments
      - deployment
      - focusAreas
      - id
      - industries
      - listed
      - name
      - organizationId
      - organizationName
      - slug
      - status
      - updatedAt
      - version
    SolutionSummary:
      type: object
      description: One solution in a list of its organization or of the operators.
      properties:
        deploymentsAwaitingReview:
          type: integer
          format: int32
          description: How many of its customer deployments wait for review.
        id:
          type: string
          format: uuid
        listed:
          type: boolean
        maturity:
          type:
          - string
          - "null"
          enum:
          - idea
          - prototype
          - pilot
          - production
          - scaled
        name:
          type: string
        organizationName:
          type: string
        slug:
          type: string
        status:
          type: string
          enum:
          - draft
          - submitted
          - approved
          - rejected
        submittedAt:
          type:
          - string
          - "null"
          format: date-time
        summary:
          type:
          - string
          - "null"
        updatedAt:
          type: string
          format: date-time
      required:
      - deploymentsAwaitingReview
      - id
      - listed
      - name
      - organizationName
      - slug
      - status
      - updatedAt
...
  securitySchemes:
    session:
      description: "The session cookie set by signing in. Requests that change state\
        \ also send `X-BeyondPilot-CSRF: 1`."
      in: cookie
      name: BEYONDPILOT_SESSION
      type: apiKey